+++ app/src/main/java/ru/netology/nmedia/domain/post/attachments/UrlAttachment.kt
@@ -5,7 +5,8 @@ import java.util.*
 data class UrlAttachment(val url: Url) : Attachment("url")
 
 data class Url(
-    val id: Int, val title: String, // Заголовок ссылки
+    val id: Long,
+    val title: String, // Заголовок ссылки
     val caption: String, // подпись ссылки
     val description: String, // описание ссылки
     val date: Date, // дата добавления
